Update DFU_71_Test_v1.0 .py
This Python script adds two buttons to the GUI. One for putting the RN4871/BM71 into CFG mode and the other for putting the RN4871/BM71 into APP mode. These buttons will be needed in order to use the BM71 tools such as the isupdate tool for updating the firmware.